pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pystructs3"
version = "0.0.8"
requires-python = ">=3.8"
dependencies = [
    'pyderive3>=0.0.7',
    'typing_extensions>=4.7.1'
]
authors = [
    {name = "Andrew C Scott", email = "imgurbot12@gmail.com"},
]
description = "Dataclass-Like Serialization Helpers for More Complex Data-Types"
readme = {file = "README.md", content-type = "text/markdown"}
license = { file = 'LICENSE' }
keywords = ["struct", "dataclass", "parser", "serialization"]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.urls]
Repository = "https://github.com/imgurbot12/pystructs"
